Make global offset unaffected by rate
The global offset is based on the latency between requesting to play a
sound and the audio device actually playing that sound. It is not
affected by the sound slowing down or speeding up, unlike local offset
(which controls the relation between the sound and the notes). I carried
out tests with large audio delay and global offset (80 ms) which confirm
that global offset should be unaffected by rate. The tests involved
setting my global offset correctly, then listening on autoplay with
hitsounds to a map on 1.0x, 0.5x and 2.0x rate. Before this change, 1.0x
would sound perfectly in sync, but on 0.5x and 2.0x the hitsounds would
be way off in differing directions. After this change they are always in
sync.