Use DesiredPoolStartingIndex in more places
This fixes the poolable scroll container sometimes not pooling enough
objects leading to the displayed objects not refreshing. Particularly
this happened when a map selection screen was opened with the 6th map
from the bottom selected: due to an outdated constant used in removed
code, the pool (of size 12) would initialize only 11 objects.

In general, in all such computations DesiredPoolStartingIndex should be
used as it always does the correct computation.