Rewrite resnap code
Now changes the time directly instead of making a deepclone.
It only resnaps the passed instead of all in the beatmap, which
makes it viable for use with only selected notes and plugin use.
It also "resnaps" notes that don't necessarily exist in the map,
which could be useful for pattern generating plugins.