ConfigManager: use _gameDirectory when writing quaver.cfg
quaver.cfg is read once at Quaver startup from _gameDirectory. Then GameDirectory might be overriden from the config property. This means, however, that setting changes would be written into quaver.cfg at the new GameDirectory location, which is never read from.

One possible fix is to change quaver.cfg reading code to check if GameDirectory is different from _gameDirectory and, if so, re-load the file from there. This, however, needs to be done recursively and with checks for GameDirectory loops, as one quaver.cfg can point to another GameDirectory, quaver.cfg there can point to a third GameDirectory, and so on. A simpler solution is to just write to _gameDirectory, which is what this commit does.