MapDatabaseCache: Double-check we're not removing existing path
I believe I've seen failure cases where people download the same map
twice in a multiplayer lobby, then the second import fails and deletes
the original file. The millisecond time has a 1/1000 chance to match,
which is not so infrequent as to literally never happen to anyone.

So, add a check: if the path we're about to delete matches the existing
map path, don't actually delete it.