Add joystick support for gameplay inputs
This commit contains two changes:
1. XnaKeyHelper is moving to Wobble.Helpers, so it's deleted from here
   and a few imports are added.
2. Several places which used to deal with Keys now deal with GenericKey.

The net result is that it's now possible to bind joystick buttons to
gameplay inputs in addition to keyboard keys. The configuration format
for keyboard keys is preserved, while joystick keys are written with the
GP_ prefix, consistent with how they are displayed in the options.

This commit changes just the gameplay inputs leaving all other inputs as
is for now.